# docs/swagger.yaml
definitions:
...
  dto.CryptoWalletType:
    enum:
    - metamask
    - walletconnect
    - coinbase
    - phantom
    - trust
    - ledger
    type: string
    x-enum-varnames:
    - WalletMetaMask
    - WalletWalletConnect
    - WalletCoinbase
    - WalletPhantom
    - WalletTrust
    - WalletLedger
...
  dto.WalletChallengeRequest:
    properties:
      wallet_address:
        maxLength: 255
        minLength: 42
        type: string
      wallet_type:
        allOf:
        - $ref: '#/definitions/dto.CryptoWalletType'
        enum:
        - metamask
        - walletconnect
        - coinbase
        - phantom
        - trust
        - ledger
    required:
    - wallet_address
    - wallet_type
    type: object
  dto.WalletConnectionRequest:
    properties:
      wallet_address:
        maxLength: 255
        minLength: 42
        type: string
      wallet_chain:
        type: string
      wallet_icon_url:
        type: string
      wallet_name:
        type: string
      wallet_type:
        allOf:
        - $ref: '#/definitions/dto.CryptoWalletType'
        enum:
        - metamask
        - walletconnect
        - coinbase
        - phantom
        - trust
        - ledger
    required:
    - wallet_address
    - wallet_chain
    - wallet_type
    type: object
  dto.WalletLoginRequest:
    properties:
      message:
        type: string
      nonce:
        type: string
      signature:
        type: string
      wallet_address:
        maxLength: 255
        minLength: 42
        type: string
      wallet_type:
        allOf:
        - $ref: '#/definitions/dto.CryptoWalletType'
        enum:
        - metamask
        - walletconnect
        - coinbase
        - phantom
        - trust
        - ledger
    required:
    - message
    - nonce
    - signature
    - wallet_address
    - wallet_type
    type: object
  dto.WalletVerificationRequest:
    properties:
      message:
        type: string
      nonce:
        type: string
      signature:
        type: string
      wallet_address:
        maxLength: 255
        minLength: 42
        type: string
      wallet_type:
        allOf:
        - $ref: '#/definitions/dto.CryptoWalletType'
        enum:
        - metamask
        - walletconnect
        - coinbase
        - phantom
        - trust
        - ledger
    required:
    - message
    - nonce
    - signature
    - wallet_address
    - wallet_type
    type: object
  response.ErrorResponse:
    properties:
      code:
        description: Code is the error code. It is not status code
        type: integer
      description:
        description: Description is the error description.
        type: string
      field_error:
        description: FieldError is the error detail for each field, if available that
          is.
        items:
          $ref: '#/definitions/response.FieldError'
        type: array
      message:
        description: Message is the error message.
        type: string
      stack_trace:
        description: |-
          StackTrace is the stack trace of the error.
          It is only returned for debugging
        type: string
    type: object
  response.FieldError:
    properties:
      description:
        description: Description is the error description for this field.
        type: string
      name:
        description: Name is the name of the field that caused the error.
        type: string
    type: object
info:
  contact: {}
paths:
...
  /api/wallet/challenge:
    post:
      consumes:
      - application/json
      description: Create a verification challenge for wallet authentication
      parameters:
      - description: Wallet Challenge Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.WalletChallengeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Create Wallet Challenge
      tags:
      - Wallet
  /api/wallet/connect:
    post:
      consumes:
      - application/json
      description: Connect a crypto wallet to the user's account
      parameters:
      - description: Bearer <token>
        in: header
        name: Authorization
        required: true
        type: string
      - description: Wallet Connection Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.WalletConnectionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Connect Wallet
      tags:
      - Wallet
...
  /api/wallet/disconnect/{connection_id}:
    delete:
      consumes:
      - application/json
      description: Disconnect a crypto wallet from the user's account
      parameters:
      - description: Bearer <token>
        in: header
        name: Authorization
        required: true
        type: string
      - description: Wallet Connection ID
        in: path
        name: connection_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Disconnect Wallet
      tags:
      - Wallet
  /api/wallet/list:
    get:
      consumes:
      - application/json
      description: Get all crypto wallets connected to the user's account
      parameters:
      - description: Bearer <token>
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Get User Wallets
      tags:
      - Wallet
  /api/wallet/login:
    post:
      consumes:
      - application/json
      description: Authenticate a user using their crypto wallet
      parameters:
      - description: Wallet Login Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.WalletLoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Login with Wallet
      tags:
      - Wallet
  /api/wallet/verify:
    post:
      consumes:
      - application/json
      description: Verify a wallet challenge signature
      parameters:
      - description: Wallet Verification Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.WalletVerificationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Verify Wallet Challenge
      tags:
      - Wallet
...
swagger: "2.0"